+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 # v0.3.0
 
+- feat: Added upstreamProvider resources
+- On authentication_code_flow, the request context is now saved in the session, instead of the browser.
+- Update fosite library, sync with hydra tag 'v25.4.0'
+
+
 # v0.2.1
 
 - The JWKS base64 encoding was incorrect. Fixed.